# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KG Extraction Backend",
    "description": "接收前端上传的 PDF，提取文本并保存为 TXT 文件",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/pdf/to-txt": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Api Pdf To Txt",
        "operationId": "api_pdf_to_txt_api_pdf_to_txt_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_pdf_to_txt_api_pdf_to_txt_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/pdf/to-txt": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Pdf To Txt",
        "operationId": "upload_pdf_to_txt_upload_pdf_to_txt_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_pdf_to_txt_upload_pdf_to_txt_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_pdf_to_txt_api_pdf_to_txt_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types"
          },
          "relations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_pdf_to_txt_api_pdf_to_txt_post"
      },
...
      "Body_upload_pdf_to_txt_upload_pdf_to_txt_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types"
          },
          "relations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_pdf_to_txt_upload_pdf_to_txt_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}